wpChanger.sh: options can be set on the commandline
the -n disappeared. If the image should be set immediately -s must be
set.
The directory for the source images can now be set with -d.
The histfile can be set with the -h option.
The resolution can be set with the -r option.
The result imagefile can be set with -o.